Switch score back to type short and changed the assignment code to be SHRT_MIN + 100
git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@931 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ sand/src/matrix.h
@@ -9,7 +9,7 @@ See the file COPYING for details.
 
 struct cell
 {
-	int score;
+	short score;
 	short traceback;
 };
